pyproject.toml
[tool.poetry]
name = "tbank"
version = "0.1.0"
description = "it purple hack, tbank case"
authors = ["Andrew <andrew.baevski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.1.1"
numpy = "2.2.3"
pandas = "2.2.3"
aiofiles = "24.1.0"
aiogram = "3.18.0"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.11.13"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
attrs = "25.2.0"
certifi = "2025.1.31"
charset-normalizer = "3.4.1"
frozenlist = "1.5.0"
idna = "3.10"
magic-filter = "1.0.12"
multidict = "6.1.0"
propcache = "0.3.0"
pydantic = "2.10.6"
pydantic-core = "2.27.2"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2025.1"
requests = "2.32.3"
six = "1.17.0"
typing-extensions = "4.12.2"
tzdata = "2025.1"
urllib3 = "2.3.0"
yarl = "1.18.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
